server/requirements.txt
Django==5.0.4
psycopg2-binary==2.9.6
gunicorn==21.2.0
djangorestframework=3.15.2